requirements.txt
Django>=3.0,<4.0
django-extensions==3.1.3
psycopg2-binary>=2.8
djangorestframework==3.12.4
Faker==8.14.0
ipdb==0.13.9
requests==2.26.0